requirements.txt
Django==3.0.8
djangorestframework==3.11.0
django-environ==0.4.5
